pyace: - upd pace_activeset: add --error-force-threshold (default auto) and --all-atoms (default - False) - add activeexploration.py - bugfix singular matrix for active set calculation
ace and ace-evaluator:
- upd multiarrays
- reduce gcc warnings, -Wvla

+++ lib/ace/ace-evaluator/src/ace-evaluator/ace_types.h
@@ -31,7 +31,7 @@
 #define ACE_TYPES_H
 
 typedef signed char RANK_TYPE;
-typedef int SPECIES_TYPE;
+typedef short int SPECIES_TYPE;
 typedef short int NS_TYPE;
 typedef short int LS_TYPE;
